+++ lib/client.dart
@@ -1,1 +1,3 @@
 export 'package:json_api/src/client/client.dart';
+export 'package:json_api/src/identifier.dart';
+export 'package:json_api/src/resource.dart';