+++ lib/src/client/client.dart
@@ -4,12 +4,14 @@ import 'dart:convert';
 import 'package:http/http.dart' as http;
 import 'package:json_api/src/client/response.dart';
 import 'package:json_api/src/client/status_code.dart';
+import 'package:json_api/src/document/collection_document.dart';
+import 'package:json_api/src/document/document.dart';
+import 'package:json_api/src/document/error_document.dart';
+import 'package:json_api/src/document/relationship.dart';
+import 'package:json_api/src/document/resource_document.dart';
+import 'package:json_api/src/document/resource_object.dart';
 import 'package:json_api/src/nullable.dart';
-import 'package:json_api/src/transport/collection_document.dart';
-import 'package:json_api/src/transport/document.dart';
-import 'package:json_api/src/transport/error_document.dart';
-import 'package:json_api/src/transport/relationship.dart';
-import 'package:json_api/src/transport/resource_document.dart';
+import 'package:json_api/src/resource.dart';
 
 typedef D ResponseParser<D extends Document>(Object j);
 
@@ -56,17 +58,13 @@ class JsonApiClient {
           {Map<String, String> headers}) =>
       _get(Relationship.fromJson, uri, headers);
 
-//  /// Creates a new resource. The resource will be added to a collection
-//  /// according to its type.
-//  Future<Response<ResourceDocument>> createResource(Uri uri, Resource resource,
-//          {Map<String, String> headers}) =>
-//      _post(
-//          ResourceDocument.fromJson,
-//          uri,
-//          ResourceDocument(ResourceObject(resource.type, resource.id,
-//              attributes: resource.attributes)),
-//          headers);
-//
+  /// Creates a new resource. The resource will be added to a collection
+  /// according to its type.
+  Future<Response<ResourceDocument>> createResource(Uri uri, Resource resource,
+          {Map<String, String> headers}) =>
+      _post(ResourceDocument.fromJson, uri,
+          ResourceDocument(ResourceObject.fromResource(resource)), headers);
+
 //  /// Adds the [identifiers] to a to-many relationship identified by [uri]
 //  Future<Response<ToMany>> addToMany(Uri uri, Iterable<Identifier> identifiers,
 //          {Map<String, String> headers}) =>
@@ -91,19 +89,19 @@ class JsonApiClient {
                 ..addAll(headers ?? {})
                 ..addAll({'Accept': contentType})));
 
-//  Future<Response<D>> _post<D extends Document>(ResponseParser<D> parse, uri,
-//          Document document, Map<String, String> headers) =>
-//      _call(
-//          parse,
-//          (_) => _.post(uri,
-//              body: json.encode(document),
-//              headers: {}
-//                ..addAll(headers ?? {})
-//                ..addAll({
-//                  'Accept': contentType,
-//                  'Content-Type': contentType,
-//                })));
-//
+  Future<Response<D>> _post<D extends Document>(ResponseParser<D> parse, uri,
+          Document document, Map<String, String> headers) =>
+      _call(
+          parse,
+          (_) => _.post(uri,
+              body: json.encode(document),
+              headers: {}
+                ..addAll(headers ?? {})
+                ..addAll({
+                  'Accept': contentType,
+                  'Content-Type': contentType,
+                })));
+
 //  Future<Response<D>> _patch<D extends Document>(ResponseParser<D> parse, uri,
 //          Document document, Map<String, String> headers) =>
 //      _call(