+++ lib/src/server/response.dart
@@ -1,17 +1,22 @@
 import 'dart:convert';
 
-import 'package:json_api/src/transport/error_document.dart';
-import 'package:json_api/src/transport/error_object.dart';
+import 'package:json_api/src/document/document.dart';
+import 'package:json_api/src/document/error_document.dart';
+import 'package:json_api/src/nullable.dart';
 
 class ServerResponse {
   final String body;
   final int status;
+  final headers = <String, String>{};
 
-  ServerResponse(this.status, {this.body});
+  ServerResponse(this.status, [Document document])
+      : body = nullable(json.encode)(document);
 
-  ServerResponse.ok([Object doc])
-      : this(200, body: doc != null ? json.encode(doc) : null);
+  ServerResponse.ok(Document document) : this(200, document);
 
-  ServerResponse.notFound({List<ErrorObject> errors = const []})
-      : this(404, body: json.encode(ErrorDocument(errors)));
+  ServerResponse.notFound(ErrorDocument document) : this(404, document);
+
+  ServerResponse.created(Document document) : this(201, document);
+
+  ServerResponse.noContent() : this(204);
 }